refactor: Consolidate pkg/platform/quantity into shared/platform/quantity
Move the quantity package from pkg/platform/ to shared/platform/ to
establish a single canonical location for platform utilities. The pkg/
directory contained only the quantity package, while shared/platform/
already housed all other platform packages (audit, auth, db, kafka, etc).

This eliminates the duplicate platform package directory and updates all
import paths across 6 services, CLI tools, load tests, and contract tests.
Documentation references in ADR-0013 and the Universal Asset System PRD
are also updated to reflect the new location.
