+++ services/financial-accounting/domain/quantity.go
@@ -6,7 +6,7 @@
 package domain
 
 import (
-	"github.com/meridianhub/meridian/pkg/platform/quantity"
+	"github.com/meridianhub/meridian/shared/platform/quantity"
 	"github.com/shopspring/decimal"
 )
