+++ shared/platform/quantity/attribute_bag.go
