+++ shared/platform/quantity/instrument.go
