+++ shared/platform/quantity/interfaces.go
