+++ shared/platform/quantity/rate.go
