+++ shared/platform/quantity/rate_test.go
